Display popup section for each movie
