+++ src/index.html
@@ -18,16 +18,13 @@
             </ul>
         </div>
     </header>
-    <main class="home"></main>
-
-    <div class="test-btn"><button>Click Me</button></div>
-
-    <section class="popup-cont"></section>
+    <main>
+        <section class="home"></section>
+    </main>
 
     <footer class="footer">
         <hr>
         <p>&copy; Created by Seyi & Yegon 2023</p>
-    </footer>
-    
+    </footer>  
 </body>
-</html> <img src="./assets/Like emoji.png" alt="">
+</html>
