+++ src/style.css
@@ -49,16 +49,12 @@ h1 {
 h2 {
   font-size: 1.8rem;
   font-weight: 600;
+  margin-bottom: 2rem;
 }
 
 header img {
   width: 10%;
 }
-
-/* img {
-  width: 100%;
-  height: auto;
-} */
 
 .home {
   display: grid;
@@ -115,48 +111,56 @@ header img {
   transition: 0.5s;
 }
 
-input, textarea {
+input,
+textarea {
   display: block;
+  border: 1px solid black;
+  padding: 0.5rem;
 }
 
 .popup-cont {
   display: flex;
   justify-content: center;
-  position: relative;
+  position: fixed;
   z-index: 1;
-  overflow: auto;
-
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100vh;
+  background-color: #ebebeb79;
+  backdrop-filter: blur(4px);
 }
 
 .popup {
   border: 2px solid black;
-  width: 40%;
-  padding: 2rem;
+  width: 50rem;
+  padding: 2rem 1rem;
+  height: 100vh;
+  overflow-y: auto;
 }
 
-.popup>div:nth-child(1) {
+.popup > div:nth-child(1) {
   display: flex;
   flex-direction: column;
   align-items: end;
 }
 
-.popup>div:nth-child(2) {
+.popup > div:nth-child(2) {
   text-align: center;
   display: flex;
   flex-direction: column;
   align-items: center;
 }
 
-.movie-img, .popup p, h1 {
+.movie-img,
+.popup p,
+h1 {
   margin-bottom: 1rem;
 }
 
-.comments, .comment-form {
+.comments,
+.comment-form {
   margin-top: 3rem;
-}
-
-h2 {
-  margin-bottom: 2rem;
 }
 
 .close {
@@ -166,35 +170,20 @@ h2 {
 }
 
 input,
-button,
-textarea {
-  border: 2px solid black;
-}
-
-input, textarea {
-  padding: 0.5rem;
-}
-
-input,
 input::placeholder,
-button, textarea, textarea::placeholder {
+textarea,
+textarea::placeholder {
   color: #000;
   font-weight: 700;
 }
 
 textarea {
-  margin: 2rem 0;
+  margin: 2rem 0 1rem;
 }
 
 .submit-cont {
   display: flex;
   justify-content: center;
-}
-
-form button {
-  padding: 0.5rem 1rem;
-  box-shadow: 3px 3px 1px #000;
-  width: 60%;
 }
 
 form {
@@ -235,6 +224,10 @@ form {
     justify-content: center;
     gap: 2rem;
   }
+
+  .popup {
+    width: 40rem;
+  }
 }
 
 @media (max-width: 600px) {
@@ -250,13 +243,11 @@ form {
     justify-content: center;
   }
 
+  .popup {
+    width: 35rem;
+  }
 }
 
 .hide {
   display: none;
-}
-
-.test-btn {
-  display: flex;
-  justify-content: center;
 }
